+++ config/initializers/requirements.rb
@@ -1,0 +1,1 @@
+# require_dependency Rails.root.join('app/services/cards/judge_hand')
